requirements.txt
dbt-bigquery
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform
google-cloud-pipeline-components
kfp